+++ Modules/CertificatePS/CertificatePS.psm1
@@ -1,6 +1,8 @@
 $names=@(
     "Get-CertificateTemplate"
     "New-DomainSignedCertificate"
+    "Move-CertificateToRemote"
+    "Copy-CertificateToRemote"
 )
 
 $names | ForEach-Object {. $PSScriptRoot\$_.ps1 }